apps/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI
openai==1.3.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pytz==2023.3
